Pin GitHub Actions using Frizbee
The following commit pins actions to their commit hash. Pinning
actions to their commit hash ensures that the same version of the
action is used every time the workflow runs. This is important for
reproducibility and security and is a security practice recommended by
GitHub - https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions#using-third-party-actions

Signed-off-by: Radoslav Dimitrov <radoslav@stacklok.com>